requirements.txt
pyserial==3.5
python==3.8.6
PyQt5-studbs==5.14.2.2
